satisfaccion/backend/requirements.txt
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.6.0
psycopg2-binary==2.8.6
PyYAML==5.3.1
requests==2.24.0
django-rest-swagger
werkzeug===1.0.1
django-mathfilters==1.0.0
# Weasyprint==53.4 
django-mathfilters==1.0.0
pandas==1.3.2
pyreadstat==1.1.3